It's fugly but it works !
I'm new to Rust, ok ?
If you think this is the worst Rust code you've seen so far.
SURE !

Still, it WORKS ! The converter works and can convert the
textures inside a GLB file to DDS (BC7 only for the moment).

Tested it with my VRCHAT GLB Loader and it worked fine !

Signed-off-by: Voyage <voyage@dontcontactmebymail.com>