Adapt the main converter method
So the idea is to return a Vec<u8> from the main converter,
so that I can export it to WASM and exploit it in a web page.

That also allows me to provide multiple interfaces to the
main method.

Also the "output" path of the main script is now used correctly.

Signed-off-by: Voyage <voyage@dontcontactmebymail.com>